audio_pipeline/requirements.txt
audiocraft>=1.3.0
torch>=2.2.0
torchaudio>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
soundfile>=0.12.1
librosa>=0.10.2.post1
pydub>=0.25.1
PyYAML>=6.0.2
tqdm>=4.66.0
